fix(soilcom): align bin size and capacity with decimal precision

+++ case_studies/soilcom/serializers.py
@@ -93,8 +93,18 @@ class CollectionModelSerializer(FieldLabelModelSerializer):
     )
     frequency = serializers.StringRelatedField()
     fee_system = serializers.StringRelatedField()
-    min_bin_size = serializers.IntegerField(required=False, allow_null=True)
-    required_bin_capacity = serializers.IntegerField(required=False, allow_null=True)
+    min_bin_size = serializers.DecimalField(
+        max_digits=8,
+        decimal_places=1,
+        required=False,
+        allow_null=True,
+    )
+    required_bin_capacity = serializers.DecimalField(
+        max_digits=8,
+        decimal_places=1,
+        required=False,
+        allow_null=True,
+    )
     required_bin_capacity_reference = serializers.CharField(
         required=False, allow_null=True
     )
@@ -190,8 +200,18 @@ class CollectionFlatSerializer(serializers.ModelSerializer):
         source="fee_system.name", label="Fee system"
     )
     frequency = serializers.StringRelatedField(label="Frequency")
-    min_bin_size = serializers.IntegerField(required=False, allow_null=True)
-    required_bin_capacity = serializers.IntegerField(required=False, allow_null=True)
+    min_bin_size = serializers.DecimalField(
+        max_digits=8,
+        decimal_places=1,
+        required=False,
+        allow_null=True,
+    )
+    required_bin_capacity = serializers.DecimalField(
+        max_digits=8,
+        decimal_places=1,
+        required=False,
+        allow_null=True,
+    )
     required_bin_capacity_reference = serializers.CharField(
         required=False, allow_null=True
     )
